Support the creation of X.509 certificates.
This is actually a little hack to turn the pkcs#10 generation code
into a full X.509 certificate generator.  Both are very similar and
with the help of 3 new functions and a few lines of code this could be
achieved.

Gpgsm (GnuPG) will be changed to allow the creation of self-signed
certificates; use that code as an example on how to create
certificates.