Add a new DER builder for internal use.
* src/der-builder.c: New.
* src/der-builder.h: New.
* src/util.c (_ksba_reallocarray): New.
--

Note that _ksba_reallocarray is the same as gpgrt_reallocarray but
because we do not want to force the use of a newer libgpg-error (aka
gpgrt) with Libksba we use a copy here.  Eventually we will make gpgrt
a mandatory dependency and then we can drop all the duplicate stuff.