Allow for NDEF list of certs and CRLs in CMS.
* src/cms-parser.c (_ksba_cms_parse_signed_data_part_2): Fix endtag
detection.

* tests/t-cms-parser.c (main): Allow several files on the command line
and add more files to the default invocation.
--

This fixes a long standing bug: TYPE_NULL was used instead of
TYPE_NONE.  However all messages received until now didn't used NDEF
encoding for the certificates so that code path was never tested or
didn't exhibit itself because another end of SET condition kicked in.

The problem was exhibited by a test message from a CryptoVision tool
which used a long list of certificates along with CRLs.  Unfortunately
we can't put a copy of that message into the test suite and that tool
is not easily available.

Signed-off-by: Werner Koch <wk@gnupg.org>